main/pyproject.toml
[tool.poetry]
name = 'space battle'
version = '0.1.0'
description = ''
authors = []
readme = 'README.md'

[tool.poetry.dependencies]
python = '^3.11'
zope-interface = "^7.1.1"
loguru = "^0.7.3"


[tool.poetry.group.dev.dependencies]
ruff = '^0.7.1'
pytest = '^8.3.3'
pytest-cov = '^5.0.0'
mypy = '^1.13.0'
coverage-badge = '^1.1.2'


[tool.pytest.ini_options]
pythonpath = ['src']

[tool.mypy]
python_version = '3.11'
show_error_codes = true
exclude = ['tests', '.venv']

[tool.ruff]
target-version = 'py311'
line-length = 120
exclude = ['.mypy_cache', '.pytest_cache', '.ruff_cache', '.venv']

[tool.ruff.lint]
select = [
    'E',    # pycodestyle
    'F',    # Pyflakes
    'UP',   # pyupgrade
    'B',    # flake8-bugbear
    'SIM',  # flake8-simplify
    'I',    # isort
]

[tool.ruff.format]
quote-style = 'single'

[build-system]
requires = ['poetry-core']
build-backend = 'poetry.core.masonry.api'
